fixed mc auto grading problem
git-svn-id: https://source.sakaiproject.org/contrib/muse/mneme/trunk@13384 fdecad78-55fc-0310-b1b2-d7d25cf747c9

+++ mneme-impl/impl/src/java/org/muse/mneme/impl/SubmissionImpl.java
@@ -766,7 +766,7 @@ public class SubmissionImpl implements Submission
 			Float score = answer.getTotalScore();
 			if (score != null)
 			{
-				total += score;
+				total += score.floatValue();
 			}
 		}
